+++ internal/pkg/client/oci/pull.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2020, Control Command Inc. All rights reserved.
-// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2023, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -12,11 +12,8 @@ import (
 	"os"
 
 	ocitypes "github.com/containers/image/v5/types"
-	"github.com/sylabs/singularity/internal/pkg/build"
-	"github.com/sylabs/singularity/internal/pkg/build/oci"
 	"github.com/sylabs/singularity/internal/pkg/cache"
 	"github.com/sylabs/singularity/internal/pkg/util/fs"
-	buildtypes "github.com/sylabs/singularity/pkg/build/types"
 	"github.com/sylabs/singularity/pkg/syfs"
 	"github.com/sylabs/singularity/pkg/sylog"
 	useragent "github.com/sylabs/singularity/pkg/util/user-agent"
@@ -30,8 +27,8 @@ type PullOptions struct {
 	NoCleanUp  bool
 }
 
-// pull will build a SIF image into the cache if directTo="", or a specific file if directTo is set.
-func pull(ctx context.Context, imgCache *cache.Handle, directTo, pullFrom string, opts PullOptions) (imagePath string, err error) {
+// sysCtx provides authentication and tempDir config for containers/image OCI operations
+func sysCtx(opts PullOptions) *ocitypes.SystemContext {
 	// DockerInsecureSkipTLSVerify is set only if --no-https is specified to honor
 	// configuration from /etc/containers/registries.conf because DockerInsecureSkipTLSVerify
 	// can have three possible values true/false and undefined, so we left it as undefined instead
@@ -47,74 +44,7 @@ func pull(ctx context.Context, imgCache *cache.Handle, directTo, pullFrom string
 	if opts.NoHTTPS {
 		sysCtx.DockerInsecureSkipTLSVerify = ocitypes.NewOptionalBool(true)
 	}
-
-	hash, err := oci.ImageDigest(ctx, pullFrom, sysCtx)
-	if err != nil {
-		return "", fmt.Errorf("failed to get checksum for %s: %s", pullFrom, err)
-	}
-
-	if directTo != "" {
-		sylog.Infof("Converting OCI blobs to SIF format")
-		if err := convertOciToSIF(ctx, imgCache, pullFrom, directTo, opts); err != nil {
-			return "", fmt.Errorf("while building SIF from layers: %v", err)
-		}
-		imagePath = directTo
-	} else {
-
-		cacheEntry, err := imgCache.GetEntry(cache.OciTempCacheType, hash)
-		if err != nil {
-			return "", fmt.Errorf("unable to check if %v exists in cache: %v", hash, err)
-		}
-		defer cacheEntry.CleanTmp()
-		if !cacheEntry.Exists {
-			sylog.Infof("Converting OCI blobs to SIF format")
-
-			if err := convertOciToSIF(ctx, imgCache, pullFrom, cacheEntry.TmpPath, opts); err != nil {
-				return "", fmt.Errorf("while building SIF from layers: %v", err)
-			}
-
-			err = cacheEntry.Finalize()
-			if err != nil {
-				return "", err
-			}
-
-		} else {
-			sylog.Infof("Using cached SIF image")
-		}
-		imagePath = cacheEntry.Path
-	}
-
-	return imagePath, nil
-}
-
-// convertOciToSIF will convert an OCI source into a SIF using the build routines
-func convertOciToSIF(ctx context.Context, imgCache *cache.Handle, image, cachedImgPath string, opts PullOptions) error {
-	if imgCache == nil {
-		return fmt.Errorf("image cache is undefined")
-	}
-
-	b, err := build.NewBuild(
-		image,
-		build.Config{
-			Dest:      cachedImgPath,
-			Format:    "sif",
-			NoCleanUp: opts.NoCleanUp,
-			Opts: buildtypes.Options{
-				TmpDir:           opts.TmpDir,
-				NoCache:          imgCache.IsDisabled(),
-				NoTest:           true,
-				NoHTTPS:          opts.NoHTTPS,
-				DockerAuthConfig: opts.OciAuth,
-				DockerDaemonHost: opts.DockerHost,
-				ImgCache:         imgCache,
-			},
-		},
-	)
-	if err != nil {
-		return fmt.Errorf("unable to create new build: %v", err)
-	}
-
-	return b.Full(ctx)
+	return sysCtx
 }
 
 // Pull will build a SIF image to the cache or direct to a temporary file if cache is disabled
@@ -130,7 +60,7 @@ func Pull(ctx context.Context, imgCache *cache.Handle, pullFrom string, opts Pul
 		sylog.Infof("Downloading library image to tmp cache: %s", directTo)
 	}
 
-	return pull(ctx, imgCache, directTo, pullFrom, opts)
+	return pullSif(ctx, imgCache, directTo, pullFrom, opts)
 }
 
 // PullToFile will build a SIF image from the specified oci URI and place it at the specified dest
@@ -141,7 +71,7 @@ func PullToFile(ctx context.Context, imgCache *cache.Handle, pullTo, pullFrom st
 		sylog.Debugf("Cache disabled, pulling directly to: %s", directTo)
 	}
 
-	src, err := pull(ctx, imgCache, directTo, pullFrom, opts)
+	src, err := pullSif(ctx, imgCache, directTo, pullFrom, opts)
 	if err != nil {
 		return "", fmt.Errorf("error fetching image to cache: %v", err)
 	}
